Move files around
Renaming and moving files in order to reflect the current structure better.
