+++ android/local.properties
@@ -1,5 +1,5 @@
 sdk.dir=/Users/filiph/Library/Android/sdk
 flutter.sdk=/Users/filiph/dev/flutter
-flutter.buildMode=release
+flutter.buildMode=debug
 flutter.versionName=0.0.2
 flutter.versionCode=2
