+++ lib/main.dart
@@ -2,10 +2,10 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:hn_app/src/article.dart';
-import 'package:hn_app/src/hn_bloc.dart';
-import 'package:hn_app/src/loading_info.dart';
-import 'package:hn_app/src/prefs_bloc.dart';
+import 'package:hn_app/src/notifiers/hn_api.dart';
+import 'package:hn_app/src/notifiers/prefs.dart';
 import 'package:hn_app/src/widgets/headline.dart';
+import 'package:hn_app/src/widgets/loading_info.dart';
 import 'package:hn_app/src/widgets/search.dart';
 import 'package:provider/provider.dart';
 import 'package:webview_flutter/webview_flutter.dart';
