+++ lib/src/widgets/loading_info.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:hn_app/src/hn_bloc.dart';
+import 'package:hn_app/src/notifiers/hn_api.dart';
 import 'package:provider/provider.dart';
 
 class LoadingInfo extends StatefulWidget {
